# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luxe Horizon Collection OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Collection {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  isPublished: boolean;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CollectionInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
}

export interface CollectionUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  slug?: string;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  isPublished?: boolean;
}

export interface ProductImage {
  id: string;
  imagePath: string;
  isPrimary: boolean;
  sortOrder: number;
}

export type ProductGender = typeof ProductGender[keyof typeof ProductGender];


export const ProductGender = {
  men: 'men',
  women: 'women',
  unknown: 'unknown',
} as const;

export type ProductCategory = string;


export const ProductCategory = {
  clothing: 'clothing',
  footwear: 'footwear',
  watches: 'watches',
  bags: 'bags',
  accessories: 'accessories',
  other: 'other',
} as const;

export interface Product {
  id: string;
  collectionId: string;
  gender: ProductGender;
  category: ProductCategory;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  aiGender?: string | null;
  /** @nullable */
  aiCategory?: string | null;
  /** @nullable */
  aiBrand?: string | null;
  /** @nullable */
  aiConfidence?: number | null;
  reviewed: boolean;
  isActive: boolean;
  isPublished: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
  images: ProductImage[];
}

export type ProductInputGender = typeof ProductInputGender[keyof typeof ProductInputGender];


export const ProductInputGender = {
  men: 'men',
  women: 'women',
  unknown: 'unknown',
} as const;

export type ProductInputCategory = string;


export const ProductInputCategory = {
  clothing: 'clothing',
  footwear: 'footwear',
  watches: 'watches',
  bags: 'bags',
  accessories: 'accessories',
  other: 'other',
} as const;

export interface ProductInput {
  collectionId: string;
  gender: ProductInputGender;
  category: ProductInputCategory;
  /** @nullable */
  brand?: string | null;
  imagePath: string;
  /** @nullable */
  imageUrl?: string | null;
}

export type ProductUpdateGender = typeof ProductUpdateGender[keyof typeof ProductUpdateGender];


export const ProductUpdateGender = {
  men: 'men',
  women: 'women',
  unknown: 'unknown',
} as const;

export type ProductUpdateCategory = string;


export const ProductUpdateCategory = {
  clothing: 'clothing',
  footwear: 'footwear',
  watches: 'watches',
  bags: 'bags',
  accessories: 'accessories',
  other: 'other',
} as const;

export interface ProductUpdate {
  gender?: ProductUpdateGender;
  category?: ProductUpdateCategory;
  /** @nullable */
  brand?: string | null;
  reviewed?: boolean;
  isActive?: boolean;
  isPublished?: boolean;
  sortOrder?: number;
}

/**
 * @nullable
 */
export type BulkProductUpdateGender = typeof BulkProductUpdateGender[keyof typeof BulkProductUpdateGender] | null;


export const BulkProductUpdateGender = {
  men: 'men',
  women: 'women',
  unknown: 'unknown',
} as const;

/**
 * @nullable
 */
export type BulkProductUpdateCategory = string | null;


export const BulkProductUpdateCategory = {
  clothing: 'clothing',
  footwear: 'footwear',
  watches: 'watches',
  bags: 'bags',
  accessories: 'accessories',
  other: 'other',
} as const;

export interface BulkProductUpdate {
  /** @minItems 1 */
  productIds: string[];
  /** @nullable */
  gender?: BulkProductUpdateGender;
  /** @nullable */
  category?: BulkProductUpdateCategory;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  reviewed?: boolean | null;
  /** @nullable */
  isPublished?: boolean | null;
}

export type UploadBatchBatchHint = typeof UploadBatchBatchHint[keyof typeof UploadBatchBatchHint];


export const UploadBatchBatchHint = {
  mixed: 'mixed',
  men: 'men',
  women: 'women',
} as const;

export type UploadBatchImagesItem = {
  imagePath: string;
  /** @nullable */
  imageUrl?: string | null;
};

export interface UploadBatch {
  collectionId: string;
  batchHint: UploadBatchBatchHint;
  category?: string;
  /** @minItems 1 */
  images: UploadBatchImagesItem[];
}

export type DashboardSummaryCategories = {[key: string]: number};

export interface DashboardSummary {
  collection: Collection;
  totalUploaded: number;
  men: number;
  women: number;
  unknown: number;
  needsReview: number;
  published: number;
  categories: DashboardSummaryCategories;
}

export interface CatalogueResult {
  collection: Collection;
  products: Product[];
  availableBrands: string[];
}

export interface PdfRequest {
  title: string;
  /** @minItems 1 */
  productIds: string[];
}

export type PdfResultStatus = typeof PdfResultStatus[keyof typeof PdfResultStatus];


export const PdfResultStatus = {
  ready: 'ready',
  generating: 'generating',
} as const;

export interface PdfResult {
  status: PdfResultStatus;
  /** @nullable */
  downloadUrl?: string | null;
  title: string;
}

export interface Settings {
  whatsappNumber: string;
  businessName: string;
  tagline: string;
}

export interface SettingsUpdate {
  whatsappNumber?: string;
  businessName?: string;
  tagline?: string;
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type CollectionQueryParameter = string;

export type GenderQueryParameter = typeof GenderQueryParameter[keyof typeof GenderQueryParameter];


export const GenderQueryParameter = {
  men: 'men',
  women: 'women',
  unknown: 'unknown',
} as const;

export type CategoryQueryParameter = string;


export const CategoryQueryParameter = {
  clothing: 'clothing',
  footwear: 'footwear',
  watches: 'watches',
  bags: 'bags',
  accessories: 'accessories',
  other: 'other',
} as const;

export type BrandQueryParameter = string;

export type ReviewQueryParameter = boolean;

export type PublishedQueryParameter = boolean;

export type ListProductsParams = {
collectionId?: CollectionQueryParameter;
gender?: GenderQueryParameter;
category?: CategoryQueryParameter;
reviewed?: ReviewQueryParameter;
published?: PublishedQueryParameter;
};

export type GetCatalogueParams = {
collectionId?: CollectionQueryParameter;
gender?: GenderQueryParameter;
category?: CategoryQueryParameter;
brand?: BrandQueryParameter;
};